requirements.txt
fastapi<0.100.0
pydantic<2.0.0
uvicorn>=0.22.0
apscheduler>=3.10.1
islamic-times>=2.1.0
pychromecast>=13.0.0
pyyaml>=6.0
requests>=2.28.0
jinja2>=3.1.2
python-multipart>=0.0.6
